requirements.txt
Flask>=2.0
pandas
tabula-py
xlwt
openpyxl
werkzeug
